+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - [fix] Upgrade playwright to v1.8.0
+- [new] A `--storybook-address` option has been added. Provide a URL where your storybook server is hosted to test those stories instead of using a static build directory.
 
 ## 2.0.1 (2021-01-22)